added loading spinner for the comparison cards search bar
